Move upgrade logic to an upgrade service
